added connections

+++ backend/controllers/history_controller.py
@@ -13,15 +13,15 @@ from backend.models.histoy_file_model import HistoryFileModel
 
 class HistoryController:
     @staticmethod
-    def store_to_history(image_data: bytes, file_ext: str, user_id: str, result):
+    def store_to_history(image_data: bytes, file_ext: str, user_id: str, result, access: str):
         try:
-            filename = HistoryController._unique_filename(file_ext)
-            image_path: str = HistoryController._get_file_path(user_id, filename)
+            filename = HistoryController.unique_filename(file_ext)
+            image_path: str = HistoryController.get_file_path(user_id, filename)
 
             with open(image_path, "wb") as f:
                 f.write(image_data)
 
-            HistoryFileModel(filename, user_id, result).save_to_db()
+            HistoryFileModel(filename, user_id, result, access).save()
 
             return filename, 200
         except Exception as e:
@@ -29,9 +29,9 @@ class HistoryController:
             return {"message": "internal_server_error"}, 500
 
     @staticmethod
-    def _get_file_path(user_id: str, filename: str) -> str:
+    def get_file_path(user_id: str, filename: str) -> str:
         current_dir = str(Path(__file__).parent.parent).replace('\\', '/')
-        user_directory = f"{current_dir}/files_history/{user_id}"
+        user_directory = f"{current_dir}/file_repository/history/{user_id}"
 
         if not os.path.exists(user_directory):
             os.makedirs(user_directory, exist_ok=True)
@@ -41,7 +41,7 @@ class HistoryController:
     @staticmethod
     def get_history_data(user_id: str):
         try:
-            result = HistoryFileModel.get_user_file(user_id)
+            result = HistoryFileModel.get_user_files(user_id)
             print(result)
             return result, 200
         except Exception as e:
@@ -52,18 +52,18 @@ class HistoryController:
     def get_file(user_id: str):
         try:
             filename = request.args.get("filename")
-            file_path = HistoryController._get_file_path(user_id, filename)
+            filepath = HistoryController.get_file_path(user_id, filename)
 
-            if not os.path.exists(file_path):
+            if not os.path.exists(filepath):
                 return {"message": "file_not_found"}, 404
 
-            return send_file(file_path, mimetype=HistoryController._get_mimetype(filename))
+            return send_file(filepath, mimetype=HistoryController.mimetype(filename))
         except Exception as e:
             Logger.error(request, e)
             return {"message": "internal_server_error"}, 500
 
     @staticmethod
-    def _get_mimetype(filename: str) -> str:
+    def mimetype(filename: str) -> str:
         ext: str = filename.split(".")[-1].lower()
 
         if ext in ["png", "jpg", "jpeg"]:
@@ -79,7 +79,7 @@ class HistoryController:
             filename = request.args.get("filename")
             if filename is None:
                 return {"message": "filename_unprovided"}, 400
-            file_path: str = HistoryController._get_file_path(user_id, filename)
+            file_path: str = HistoryController.get_file_path(user_id, filename)
             HistoryFileModel.delete_file(user_id, filename)
             os.remove(file_path)
             return filename, 200
@@ -90,8 +90,9 @@ class HistoryController:
     @staticmethod
     def clear_history(user_id):
         try:
-            file_path: str = HistoryController._get_file_path(user_id, "")[:-1]
-            shutil.rmtree(file_path)
+            file_path: str = HistoryController.get_file_path(user_id, "")[:-1]
+            if os.path.exists(file_path):
+                shutil.rmtree(file_path)
             result: int = HistoryFileModel.delete_user_files(user_id)
 
             return {"deletedFilesCount": result}, 200
@@ -102,7 +103,7 @@ class HistoryController:
             return {"message": "internal_server_error"}, 500
 
     @staticmethod
-    def _unique_filename(ext: str) -> str:
+    def unique_filename(ext: str) -> str:
         num = int(time.time() * 1000)
         time_bytes: bytes = num.to_bytes((num.bit_length() + 7) // 8, byteorder='big')
         time_base64: str = base64.b64encode(time_bytes).decode('utf-8')
